fixed shadow color

+++ illumination/assets/materials/shadow/shadow_vp.glsl
@@ -1,7 +1,7 @@
 #ifndef shadow_vp
 #define shadow_vp
 
-uniform mediump mat4 mtx_light;
+uniform highp mat4 mtx_light;
 varying highp vec4 var_texcoord0_shadow;
 
 #endif
